+++ Pwinty.Client/ShippingRate.cs
@@ -9,7 +9,7 @@ namespace Pwinty.Client
     {
         public string Id { get; set; }
         public string Description { get; set; }
-        public decimal PriceGBP { get; set; }
-        public decimal PriceUSD { get; set; }
+        public int PriceGBP { get; set; }
+        public int PriceUSD { get; set; }
     }
 }
